requirements.txt
shap
optuna
torch
torchvision
opencv-python
scikit-learn
scikit-image
matplotlib
numpy
Pillow
numba
imageio